ENH: Split dashboard component for better readability
