+++ front_end/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@emotion/core": "10.0.9",
     "@emotion/styled": "10.0.9",
+    "@reduxjs/toolkit": "^1.9.7",
     "abort-controller": "^3.0.0",
     "axios": "^0.21.1",
     "bootstrap": "^5.3.2",
@@ -37,7 +38,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-i18next": "^11.8.5",
-    "react-redux": "^7.2.4",
+    "react-redux": "^8.1.3",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^4.0.3",
     "reactstrap": "^8.4.1",
