+++ front_end/src/redux/store.js
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit'
+
+import pageReducer from './inventory_reducer'
+
+export default configureStore({
+  reducer: {
+    page: pageReducer
+  },
+})
